autojobagent/requirements.txt
streamlit>=1.30
requests>=2.31
beautifulsoup4>=4.12
sqlalchemy>=2.0
python-dotenv>=1.0
pyyaml>=6.0
sentence-transformers>=5.0
transformers>=5.0
torch>=2.0
